refactor(UI): ♻️ Relocated testimonials & standardized page naming.     -   Moved the portrait and testimonial quotes to the header, which     ensures key social proof is visible to users immediately upon landing.
    -   Removed redundant imagery from the Personal view, which declutters
    the interface and speeds up page loading times.

    -   Standardized page filenames to lowercase, which prevents broken
    links and ensures consistent behavior across different web servers.

    -   Updated the reference display logic, which improves the visual
    hierarchy and makes professional endorsements easier to read.

    -   Refined the navigation structure, which simplifies the user
    journey and creates a more professional brand presentation.
